Use 16-bit depth field in rendering
This makes normals look better, because we're more likely to calculate
the gradient for the right point on the solid.

+++ src/fab/tree/render.h
@@ -21,7 +21,7 @@ void render8(struct MathTree_* tree, Region region,
              uint8_t** img, volatile int* halt,
              void (*callback)());
 
-void shaded8(struct MathTree_* tree, Region region, uint8_t** depth,
+void shaded8(struct MathTree_* tree, Region region, uint16_t** depth,
              uint8_t (**out)[3], volatile int* halt,
              void (*callback)());
